Move certificate helper in as a private local class
